moving to global events for strategies
