+++ src/global.js
@@ -10,8 +10,12 @@ globalEvents.TI4 = {
     onSystemActivated: new TriggerableMulticastDelegate(),
 
     // Called when a Strategy Card is Played
-    // <(object: cardd, player:Player) => void>
+    // <(object: card, player:Player) => void>
     onStrategyCardPlayed: new TriggerableMulticastDelegate(),
+
+    // Called when a Strategy Card selection is done by a player
+    // <(object: card, player:Player) => void>
+    onStrategyCardSelectionDone: new TriggerableMulticastDelegate(),
 };
 
 require("./global/patch-infinite-container");
